fix(p5): phpstan L3 — zero errors
Climb phpstan.neon from level 2 to level 3 and resolve all resulting
findings — wrong array-shape/return-type docblocks, real int/float
precision fixes, and two by-ref-populated-array cases requiring the
compute()-style @param-out fix from L1/L2. No suppressions except one:
Smarty's own vendored TemplateBase.php declares `@var string` on
$compile_id while defaulting it to null — not ours to fix.

Wrong docblocks describing a fixed literal/tuple shape instead of the
real, mutable contract (same root cause repeated across many files —
someone described the constructor's initial value, not what the property
holds once used): tabsheet::$sheets/$selected/$uniqid, updates::$types/
$default_plugins/$default_themes, PwgXmlWriter::$_elementStack/
$_lastTagOpen, pwg_image's crop()/resize() docblocks, ImageStdParams'
$undefined_type_map (this one was actually breaking derivative.inc.php's
`$ret[$type] = $ret[$type2]` with "Invalid array key type DerivativeParams",
since the WRONG @var/@ return said DerivativeParams[] on an array that's
genuinely string=>string).

Real bugs found and fixed, not just docblock corrections:
- admin/site_update.php: $simulate read at 5 points outside the
  $_POST['submit'] gate that sets it — same defensive-init pattern as the
  L1 $counts/$basedir fix, plus simplified the now-redundant
  isset($simulate) guard.
- include/functions.inc.php: two navbar page-number loops used floor()/
  ceil() results (always float in PHP) as array keys — cast to int at the
  computation site, matching the actual "this is a page number" semantics.
- include/derivative_params.inc.php: SizingParams::compute()'s scale_size
  math used floor() results directly; cast to int where the value is a
  final pixel dimension. ImageRect's $l/$t/$r/$b properties turned out to
  genuinely need int|float (crop_h()/crop_v() do float arithmetic on them),
  so width()/height() keep their real int|float return rather than forcing
  everything through more casts than the actual crop math tolerates.
- include/template.class.php: ScriptLoader::get_footer_scripts() docblock
  claimed a flat Combinable[] but the body always returns a 2-tuple of
  Combinable[] arrays (checked do_combine()'s own real return type before
  writing the fix) — the wrong docblock was making func_get_combined_scripts()'s
  two foreach loops look like they were iterating a Combinable object.
  Combinable::$version genuinely holds string|false (false disables
  version-based cache busting, used via three separate === false / !== false
  checks elsewhere in the same file) — widened, not narrowed to match one
  call site.
- i.php: $page built up across parse_request() (global-scope mutation
  PHPStan can't trace) plus later top-level DB-lookup code; added the
  established defensive-init block (11 keys, one per assign site) rather
  than another @param-out attempt, since parse_request() takes no
  parameter to hang @param-out off of.
- include/common.inc.php: $user['id']/['email']/['theme'] read after
  `include user.inc.php`, which only ever calls build_user()/auto_login()/
  auth_key_login() — never assigns $user directly — so PHPStan can't trace
  it through the include the way it can for genuinely inline top-level
  code. Same defensive-init treatment.
- admin/include/pclzip.lib.php: privReadEndCentralDir() built its by-ref
  output array key-by-key; PHPStan's @param-out checks a by-ref array
  shape at every intermediate write, not just the final one, so incremental
  key assignment can never satisfy an exact-shape contract. Refactored to
  one atomic array literal (real code-quality improvement, not just a
  types workaround) and gave privMerge()'s untyped $p_archive_to_add
  parameter its real PclZip type — needed for @param-out to resolve
  through `$p_archive_to_add->privReadEndCentralDir(...)` at all.

Full suite green: composer test, test:integration, test:contract,
test:browser (44 passed, 1 skipped), test:visual (32/32), lint:php,
plan-lint.

+++ admin/include/functions_upload.inc.php
@@ -1017,5 +1017,5 @@ function get_optimal_dimensions_for_representative(): array
 
     $margin_coef = 1.5;
 
-    return [$w * $margin_coef, $h * $margin_coef];
+    return [(int) ($w * $margin_coef), (int) ($h * $margin_coef)];
 }
